+++ src/controllers/category-controller.js
@@ -60,6 +60,7 @@ export const categoryController = {
   deletePlacemark: {
     handler: async function (request, h) {
       const category = await db.categoryStore.getCategoryById(request.params.id);
+      await db.reviewStore.deleteReviewsByPlacemarkId(request.params.placemarkid);
       await db.placemarkStore.deletePlacemark(request.params.placemarkid);
       return h.redirect(`/category/${category._id}`);
     },